backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.0
djangorestframework-simplejwt>=5.3
django-environ>=0.10
psycopg2-binary>=2.9
pymongo>=4.5
djongo>=1.3
redis>=5.0
celery>=5.3
boto3>=1.28
cloudinary>=1.36
